+++ .github/workflows/integration-tests.yml
@@ -67,7 +67,6 @@ jobs:
             apiml:
                 image: ghcr.io/balhar-jakub/apiml:${{ github.run_id }}-${{ github.run_number }}
                 env:
-                    APIML_SECURITY_SSL_VERIFYSSLCERTIFICATESOFSERVICES: true
                     APIML_SECURITY_SSL_NONSTRICTVERIFYSSLCERTIFICATESOFSERVICES: true
                     APIML_DISCOVERY_STATICAPIDEFINITIONSDIRECTORIES: /docker/api-defs-unknown-host
             discoverable-client-unknown:
