+++ apiml-common/src/main/java/org/zowe/apiml/product/web/HttpConfig.java
@@ -86,9 +86,6 @@ public class HttpConfig implements InitializingBean {
     @Value("${apiml.service.ssl.key-store-type:${server.ssl.keyStoreType:PKCS12}}")
     private String keyStoreType;
 
-    @Value("${apiml.security.ssl.verifySslCertificatesOfServices:true}")
-    private boolean verifySslCertificatesOfServices;
-
     @Value("${apiml.security.ssl.nonStrictVerifySslCertificatesOfServices:false}")
     private boolean nonStrictVerifySslCertificatesOfServices;
 
@@ -164,7 +161,7 @@ public class HttpConfig implements InitializingBean {
                     .protocol(protocol).enabledProtocols(supportedProtocols).cipherSuite(ciphers)
                     .trustStore(trustStorePath).trustStoreType(trustStoreType)
                     .trustStorePassword(trustStorePassword).trustStoreRequired(trustStoreRequired)
-                    .verifySslCertificatesOfServices(verifySslCertificatesOfServices)
+                    .verifySslCertificatesOfServices(true)
                     .nonStrictVerifySslCertificatesOfServices(nonStrictVerifySslCertificatesOfServices)
                     .maxConnectionsPerRoute(maxConnectionsPerRoute).maxTotalConnections(maxTotalConnections)
                     .idleConnTimeoutSeconds(idleConnTimeoutSeconds).requestConnectionTimeout(requestConnectionTimeout)
