+++ apiml-security-common/build.gradle
@@ -19,7 +19,6 @@ dependencies {
     implementation libs.nimbus.jose.jwt // Parsing
 
     testImplementation libs.spring.boot.starter.test
-    testImplementation libs.reactor.test
     testImplementation(testFixtures(project(":apiml-common")))
 
     compileOnly libs.lombok
