+++ caching-service-package/src/main/resources/bin/start.sh
@@ -112,7 +112,6 @@ _BPX_JOBNAME=${ZWE_zowe_job_prefix}${CACHING_CODE} ${JAVA_BIN_DIR}java \
   -Dapiml.service.hostname=${ZWE_haInstance_hostname:-localhost} \
   -Dapiml.service.discoveryServiceUrls=${ZWE_DISCOVERY_SERVICES_LIST} \
   -Dapiml.service.customMetadata.apiml.gatewayPort=${ZWE_components_gateway_port:-7554} \
-  -Dapiml.service.ssl.verifySslCertificatesOfServices=${verifySslCertificatesOfServices:-false} \
   -Dapiml.service.ssl.nonStrictVerifySslCertificatesOfServices=${nonStrictVerifySslCertificatesOfServices:-false} \
   -Dapiml.service.ssl.enabled-protocols=${ZWE_configs_apiml_service_ssl_enabled_protocols:-${client_enabled_protocols}} \
   -Dapiml.service.ssl.ciphers=${ZWE_configs_apiml_service_ssl_ciphers:-${client_ciphers}} \
@@ -126,8 +125,6 @@ _BPX_JOBNAME=${ZWE_zowe_job_prefix}${CACHING_CODE} ${JAVA_BIN_DIR}java \
   -Dapiml.service.ssl.trust-store-password="${client_truststore_pass}" \
   -Dapiml.service.ssl.trust-store-type="${client_truststore_type}" \
   -Dapiml.security.x509.certificatesUrls=${CERTIFICATES_URLS} \
-  -Dapiml.service.http.userId=${ZWE_configs_apiml_service_http_userId:-} \
-  -Dapiml.service.http.password=${ZWE_configs_apiml_service_http_password:-} \
   -Djdk.tls.client.cipherSuites=${client_ciphers} \
   -Dserver.ssl.ciphers=${server_ciphers} \
   -Dserver.ssl.protocol=${server_protocol} \
