+++ caching-service-package/src/main/resources/schemas/caching-service-config.json
@@ -110,20 +110,6 @@
                                     "default": false
                                 }
                             }
-                        },
-                        "http": {
-                            "type": "object",
-                            "description": "Configuration for service running in HTTP mode.",
-                            "properties": {
-                                "userId": {
-                                    "type": "string",
-                                    "description": "Username used for internal authentication inside APIML. It is used only when SSL certificate verification is disabled."
-                                },
-                                "password": {
-                                    "type": "string",
-                                    "description": "Password used for internal authentication inside APIML. It is used only when SSL certificate verification is disabled."
-                                }
-                            }
                         }
                     }
                 }
