--- caching-service/src/test/java/org/zowe/apiml/caching/config/SecurityConfigTest.java
@@ -1,118 +1,0 @@
-/*
- * This program and the accompanying materials are made available under the terms of the
- * Eclipse Public License v2.0 which accompanies this distribution, and is available at
- * https://www.eclipse.org/legal/epl-v20.html
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Copyright Contributors to the Zowe Project.
- */
-
-package org.zowe.apiml.caching.config;
-
-import io.restassured.RestAssured;
-import io.restassured.http.Header;
-import org.junit.jupiter.api.*;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.test.context.TestPropertySource;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import org.zowe.apiml.caching.CachingServiceApplication;
-import org.zowe.apiml.security.common.verify.CertificateValidator;
-import org.zowe.apiml.util.config.SslContext;
-import org.zowe.apiml.util.config.SslContextConfigurer;
-
-import static io.restassured.RestAssured.given;
-
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class SecurityConfigTest {
-
-    @BeforeAll
-    static void init() throws Exception {
-        SslContext.reset();
-        RestAssured.useRelaxedHTTPSValidation();
-        SslContextConfigurer configurer = new SslContextConfigurer("password".toCharArray(), "../keystore/client_cert/client-certs.p12", "../keystore/localhost/localhost.keystore.p12");
-        SslContext.prepareSslAuthentication(configurer);
-    }
-
-    private String getUri(String hostname, int port) {
-        return String.format("%s://%s:%d/%s", "https", hostname, port, "cachingservice/api/v1/cache");
-    }
-
-    @Nested
-    @TestPropertySource(
-        properties = {
-            "apiml.service.ssl.verifySslCertificatesOfServices=true"
-        }
-    )
-    @SpringBootTest(
-        classes = CachingServiceApplication.class,
-        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT
-    )
-    class GivenEnabledSSLVerification {
-
-        @Value("${apiml.service.hostname:localhost}")
-        String hostname;
-        @LocalServerPort
-        int port;
-
-        @MockitoBean
-        private CertificateValidator certificateValidator;
-
-        private static final String MOCK_FORWARDED_CERT_HEADER = """
-            MIID7zCCAtegAwIBAgIED0TPEjANBgkqhkiG9w0BAQsFADB6MQswCQYDVQQGEwJD
-            WjEPMA0GA1UECBMGUHJhZ3VlMQ8wDQYDVQQHEwZQcmFndWUxFDASBgNVBAoTC1pv
-            d2UgU2FtcGxlMRwwGgYDVQQLExNBUEkgTWVkaWF0aW9uIExheWVyMRUwEwYDVQQD
-            Ewxab3dlIFNlcnZpY2UwHhcNMTgxMjA3MTQ1NzIyWhcNMjgxMjA0MTQ1NzIyWjB6
-            MQswCQYDVQQGEwJDWjEPMA0GA1UECBMGUHJhZ3VlMQ8wDQYDVQQHEwZQcmFndWUx
-            FDASBgNVBAoTC1pvd2UgU2FtcGxlMRwwGgYDVQQLExNBUEkgTWVkaWF0aW9uIExh
-            eWVyMRUwEwYDVQQDEwxab3dlIFNlcnZpY2UwggEiMA0GCSqGSIb3DQEBAQUAA4IB
-            DwAwggEKAoIBAQC6Orc/EJ5/t2qam1DiYU/xVbHaQrjd6uvpj2HTvOOohtFZ7/Kx
-            yMAezgB8DBR4+77qXXsdP9ngnTl/i22yGwvo7Tlz6dhnQLnks7VFr1eGGC2ks+rL
-            BJsF/RQexmONG9ddexWD8SOYoW9RRapQqETbcllxOenvzXruOEzaXhMazkK9Cg+J
-            ucNb9HcfhIM0rjLZhqG8Gc8dAtCcxF/xHlVyFQq8fr4u2p/wGmARM14iZeQltQV7
-            F3gxmw3djfcNM5S3tirPrHlZb76ZmmQEn4QiLSP198Lm+4QKAOw1dUpMf4eELO4c
-            EFUHXQUCHLWc5NztZxWW40NrDbZEjcRI5ah7AgMBAAGjfTB7MB0GA1UdJQQWMBQG
-            CCsGAQUFBwMCBggrBgEFBQcDATAOBgNVHQ8BAf8EBAMCBPAwKwYDVR0RBCQwIoIV
-            bG9jYWxob3N0LmxvY2FsZG9tYWlugglsb2NhbGhvc3QwHQYDVR0OBBYEFHL1ygBb
-            UCI/ktdk3TgQA6EJlATIMA0GCSqGSIb3DQEBCwUAA4IBAQBHALBlFf0P1TBR1MHQ
-            vXYDFAW+PiyF7zP0HcrvQTAGYhF7uJtRIamapjUdIsDVbqY0RhoFnBOu8ti2z0pW
-            djw47f3X/yj98n+J2aYcO64Ar+ovx93P01MA8+Mz1u/LwXk4pmrbUIcOEtyNu+vT
-            a0jDobC++3Zfv5Y+iD2M8L+jacSMZNCqQByhKtTkAICXg9LMccx4XLYtJ65zGP2h
-            4TEK0MMfO2G1/vUmdb3tq17zKdukj3MUS254mENCck7ioNFR0Cc9lzuSHyBrdb0x
-            M/iHeamNblckK/r1roDjhCAQz9DtmETad/o7qGNFxDTRRShRV9Lww0fFB7PaV7u/
-            VPx2
-            """.replaceAll("\\s+", "");
-
-        @BeforeEach
-        void setup() {
-            org.mockito.Mockito.when(certificateValidator.isForwardingEnabled()).thenReturn(true);
-            org.mockito.Mockito.when(certificateValidator.hasGatewayChain(org.mockito.Mockito.any())).thenReturn(true);
-        }
-
-        @Test
-        void whenNoClientCertificate_thenReturnUnauthorized() {
-            given()
-                .header(new Header("X-CS-Service-ID", "apimtst"))
-                .get(getUri(hostname, port))
-                .then()
-                .log().ifValidationFails()
-                .statusCode(HttpStatus.FORBIDDEN.value());
-        }
-
-        @Test
-        void whenClientCertificate_thenReturnOk() {
-
-            given()
-                .config(SslContext.clientCertApiml)
-                .header(new Header("X-CS-Service-ID", "apimtst"))
-                .header(new Header("Client-Cert", MOCK_FORWARDED_CERT_HEADER))
-                .get(getUri(hostname, port))
-                .then()
-                .log().ifValidationFails()
-                .statusCode(HttpStatus.OK.value());
-        }
-    }
-}
