+++ config/docker/discovery-service.yml
@@ -14,8 +14,6 @@ apiml:
         allPeersUrls: https://discovery-service:10011/eureka
     security:
         allowedDomains: discovery-service,discovery-service-2,discoverable-client,discoverable-client-2,api-catalog-services,api-catalog-services-2,caching-service,caching-service-2,caching-service-3,gateway-service,gateway-service-2,zaas-service,zaas-service-2,mock-services,mock-services-2,zowe.github.io,www.zowe.org,localhost,www.ibm.com,central-gateway-service,central-gateway-service-2
-        ssl:
-            verifySslCertificatesOfServices: true
     banner: console
 spring:
     output:
