+++ config/local-multi/apiml-1.yml
@@ -41,8 +41,6 @@ apiml:
     discovery:
         allPeersUrls: https://localhost:10011/eureka,https://localhost:10021/eureka
     security:
-        # ssl:
-        #     verifySslCertificatesOfServices: false
         x509:
             enabled: true
             certificatesUrl: https://localhost:10010/gateway/certificates
