+++ config/local/api-defs-http/zosmf-sample.yml_
@@ -1,18 +1,6 @@
 #
 # Static definition for z/OSMF
 #
-# To skip adding service certificate or CA certificate to the truststore
-# of the API Gateway you need to see following setting in the
-# config/local/gateway-service.yml:
-#
-#   apiml:
-#       security:
-#           verifySslCertificatesOfServices: false
-#
-#  Then you can access z/OSMF via the API gateway:
-#
-#   http --verify=keystore/local_ca/localca.cer GET https://localhost:10010/api/{zosmf-service-id}/zosmf/info 'X-CSRF-ZOSMF-HEADER;'
-#
 services:
     - serviceId: ibmzosmf                                         # unique lowercase ID of the service
       title: IBM z/OSMF                                           # Title of the z/OSMF service in the API catalog
