+++ config/local/discovery-service.yml
@@ -10,7 +10,6 @@ apiml:
         staticApiDefinitionsDirectories: config/local/api-defs
     security:
         allowedDomains: localhost,localhost2,www.ibm.com,zowe.github.io,www.zowe.org
-        ssl.verifySslCertificatesOfServices: true
     banner: console
 
 spring.output.ansi.enabled: always
