+++ config/local/zaas-service.yml
@@ -30,8 +30,6 @@ apiml:
             provider: zosmf
             zosmf:
                 serviceId: mockzosmf  # Replace me with the correct z/OSMF service id
-        ssl:
-            verifySslCertificatesOfServices: true
         x509:
             enabled: true
             acceptForwardedCert: true
