+++ discovery-service/src/test/resources/ext-config/zosmf-sample.yml
@@ -1,18 +1,6 @@
 #
 # Static definition for z/OSMF
 #
-# To skip adding service certificate or CA certificate to the truststore
-# of the API Gateway you need to see following setting in the
-# config/local/gateway-service.yml:
-#
-#   apiml:
-#       security:
-#           verifySslCertificatesOfServices: false
-#
-#  Then you can access z/OSMF via the API gateway:
-#
-#   http --verify=keystore/local_ca/localca.cer GET https://localhost:10010/api/{zosmf-service-id}/zosmf/info 'X-CSRF-ZOSMF-HEADER;'
-#
 services:
     - serviceId: zosmfserviceid  # unique lowercase ID of the service
       title: IBM z/OSMF
