+++ gateway-package/src/main/resources/schemas/gateway-config.json
@@ -582,20 +582,6 @@
                                     "default": false
                                 }
                             }
-                        },
-                        "http": {
-                            "type": "object",
-                            "description": "Configuration for service running in HTTP mode.",
-                            "properties": {
-                                "userId": {
-                                    "type": "string",
-                                    "description": "Username used for internal authentication inside APIML."
-                                },
-                                "password": {
-                                    "type": "string",
-                                    "description": "Password used for internal authentication inside APIML."
-                                }
-                            }
                         }
                     }
                 },
