+++ gateway-service/src/test/java/org/zowe/apiml/gateway/caching/CachingServiceClientRestTest.java
@@ -17,29 +17,22 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.test.util.ReflectionTestUtils;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.reactive.function.client.ClientRequest;
 import org.springframework.web.reactive.function.client.ClientResponse;
 import org.springframework.web.reactive.function.client.ExchangeFunction;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.zowe.apiml.gateway.caching.CachingServiceClient.ApiKeyValue;
 import org.zowe.apiml.gateway.caching.LoadBalancerCache.LoadBalancerCacheRecord;
-import org.zowe.apiml.message.log.ApimlLogger;
 import org.zowe.apiml.product.gateway.GatewayClient;
 import org.zowe.apiml.product.instance.ServiceAddress;
 import reactor.test.StepVerifier;
 
 import java.util.function.Predicate;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 import static reactor.core.publisher.Mono.empty;
@@ -53,9 +46,6 @@ class CachingServiceClientRestTest {
 
     @Mock
     private ClientResponse clientResponse;
-
-    @Mock
-    private ApimlLogger apimlLogger;
 
     private CachingServiceClientRest client;
     private WebClient webClient;
@@ -231,48 +221,6 @@ class CachingServiceClientRestTest {
                     .verifyErrorMatches(assertCachingServiceClientException(404));
             }
 
-        }
-
-    }
-
-    @Nested
-    class CachingServiceAuthorization {
-
-        @Test
-        void givenCredentials_whenSetCredentials_thenSetAuthorizationHeader() {
-            var service = new CachingServiceClientRest(null, null);
-            ReflectionTestUtils.setField(service, "cachingServiceUserId", "user");
-            ReflectionTestUtils.setField(service, "cachingServicePassword", "password");
-            ReflectionTestUtils.setField(service, "apimlLog", apimlLogger);
-
-            service.afterPropertiesSet();
-
-            // Verify that no warning is logged
-            verify(apimlLogger, times(0)).log("org.zowe.apiml.security.common.auth.missingDefaultCredentials");
-            // Verify that Basic authHeader is in the defaultHeaders map
-            var headers = (MultiValueMap<String, String>) ReflectionTestUtils.getField(service, "defaultHeaders");
-            assertEquals("Basic dXNlcjpwYXNzd29yZA==", headers.get(HttpHeaders.AUTHORIZATION).get(0));
-        }
-
-        @ParameterizedTest
-        @CsvSource({
-            ",password,,",
-            "user,,",
-            ",,"
-        })
-        void givenIncompleteCredentials_whenSetCredentials_thenDoNotSetAuthorization(String userId, String password) {
-            var service = new CachingServiceClientRest(null, null);
-            ReflectionTestUtils.setField(service, "cachingServiceUserId", userId);
-            ReflectionTestUtils.setField(service, "cachingServicePassword", password);
-            ReflectionTestUtils.setField(service, "apimlLog", apimlLogger);
-
-            service.afterPropertiesSet();
-
-            // Verify that warning is logged
-            verify(apimlLogger, times(1)).log("org.zowe.apiml.security.common.auth.missingDefaultCredentials");
-            // Verify that Basic authHeader is not in the defaultHeaders map
-            var headers = (MultiValueMap<String, String>) ReflectionTestUtils.getField(service, "defaultHeaders");
-            assertNull(headers.get(HttpHeaders.AUTHORIZATION));
         }
 
     }
