+++ integration-tests/src/test/resources/environment-configuration-attls.yml
@@ -93,7 +93,6 @@ instanceEnv:
     ZWE_CACHING_SERVICE_PORT: 10016
     ZWE_CACHING_EVICTION_STRATEGY: reject
     ZWE_CACHING_STORAGE_SIZE: 100
-    VERIFY_CERTIFICATES: true
     APIML_SSL_ENABLED: false
     EUREKA_CLIENT_REGISTRY-FETCH-INTERVAL-SECONDS: 5
     EUREKA_INSTANCE_LEASE-EXPIRATION-DURATION-IN-SECONDS: 45
