+++ zaas-package/src/main/resources/bin/start.sh
@@ -157,7 +157,7 @@ _BPX_JOBNAME=${ZWE_zowe_job_prefix}${ZAAS_CODE} ${JAVA_BIN_DIR}java \
     -Dapiml.connection.idleConnectionTimeoutSeconds=${ZWE_configs_apiml_connection_idleConnectionTimeoutSeconds:-5} \
     -Dapiml.cache.storage.location=${ZWE_zowe_workspaceDirectory}/api-mediation/${ZWE_haInstance_id:-localhost} \
     -Dapiml.logs.location=${ZWE_zowe_logDirectory} \
-    -Dapiml.security.ssl.verifySslCertificatesOfServices=${verifySslCertificatesOfServices:-true} \
+    -Dapiml.security.ssl.verifySslCertificatesOfServices=true \
     -Dapiml.security.ssl.nonStrictVerifySslCertificatesOfServices=${nonStrictVerifySslCertificatesOfServices:-false} \
     -Dapiml.security.auth.zosmf.serviceId=${ZWE_configs_apiml_security_auth_zosmf_serviceId:-${ZWE_components_gateway_apiml_security_auth_zosmf_serviceId:-ibmzosmf}} \
     -Dapiml.security.auth.provider=${ZWE_configs_apiml_security_auth_provider:-${ZWE_components_gateway_apiml_security_auth_provider:-zosmf}} \
@@ -177,8 +177,6 @@ _BPX_JOBNAME=${ZWE_zowe_job_prefix}${ZAAS_CODE} ${JAVA_BIN_DIR}java \
     -Dapiml.service.ssl.trust-store="${client_truststore_location}" \
     -Dapiml.service.ssl.trust-store-password="${client_truststore_pass}" \
     -Dapiml.service.ssl.trust-store-type="${client_truststore_type}" \
-    -Dapiml.service.http.userId=${ZWE_configs_apiml_service_http_userId:-} \
-    -Dapiml.service.http.password=${ZWE_configs_apiml_service_http_password:-} \
     -Djdk.tls.client.cipherSuites=${client_ciphers} \
     -Dserver.ssl.ciphers=${server_ciphers} \
     -Dserver.ssl.protocol=${server_protocol} \
