+++ zaas-service/src/main/java/org/zowe/apiml/zaas/cache/CachingServiceClient.java
@@ -15,20 +15,15 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.*;
 import org.springframework.web.client.HttpStatusCodeException;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
-import org.zowe.apiml.message.log.ApimlLogger;
 import org.zowe.apiml.product.gateway.GatewayClient;
 import org.zowe.apiml.product.instance.ServiceAddress;
-import org.zowe.apiml.product.logging.annotations.InjectApimlLogger;
 
-import java.util.Base64;
 import java.util.Map;
 
 
@@ -38,28 +33,16 @@ import java.util.Map;
  */
 @Slf4j
 @SuppressWarnings({"squid:S1192"}) // literals are repeating in debug logs only
-public class CachingServiceClient implements CachingClient, InitializingBean {
+public class CachingServiceClient implements CachingClient {
 
     private final GatewayClient gatewayClient;
     private final RestTemplate restTemplate;
-
-    @InjectApimlLogger
-    private final ApimlLogger apimlLog = ApimlLogger.empty();
 
     @Value("${apiml.cachingServiceClient.apiPath:/cachingservice/api/v1/cache}")
     private String CACHING_API_PATH;
 
     @Value("${apiml.cachingServiceClient.list.apiPath:/cachingservice/api/v1/cache-list/}")
     private String CACHING_LIST_API_PATH;
-
-    @Value("${apiml.service.http.userId:#{null}}")
-    private String cachingServiceUserId;
-
-    @Value("${apiml.service.http.password:#{null}}")
-    private String cachingServicePassword;
-
-    @Value("${apiml.security.ssl.verifySslCertificatesOfServices:true}")
-    private boolean verifyCertificates;
 
     private static final HttpHeaders defaultHeaders = new HttpHeaders();
 
@@ -77,18 +60,6 @@ public class CachingServiceClient implements CachingClient, InitializingBean {
             throw new IllegalStateException("RestTemplate instance cannot be null");
         }
         this.restTemplate = restTemplate;
-    }
-
-    @Override
-    public void afterPropertiesSet() {
-        if (!verifyCertificates) {
-            if (StringUtils.isEmpty(cachingServiceUserId) || StringUtils.isEmpty(cachingServicePassword)) {
-                apimlLog.log("org.zowe.apiml.security.common.auth.missingDefaultCredentials");
-            } else {
-                String basicToken = "Basic " + Base64.getEncoder().encodeToString((cachingServiceUserId + ":" + cachingServicePassword).getBytes());
-                defaultHeaders.add(HttpHeaders.AUTHORIZATION, basicToken);
-            }
-        }
     }
 
     private String getGatewayAddress() {
