+++ zaas-service/src/test/java/org/zowe/apiml/zaas/cache/CachingServiceClientTest.java
@@ -15,19 +15,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.*;
 import org.springframework.test.util.ReflectionTestUtils;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
-import org.zowe.apiml.message.log.ApimlLogger;
 import org.zowe.apiml.models.AccessTokenContainer;
 import org.zowe.apiml.product.gateway.GatewayClient;
 import org.zowe.apiml.product.instance.ServiceAddress;
@@ -257,52 +250,6 @@ class CachingServiceClientTest {
             doThrow(new RestClientException("oops")).when(restTemplate).exchange(anyString(), any(HttpMethod.class), any(HttpEntity.class), eq(String.class));
             assertThrows(CachingServiceClientException.class, () -> underTest.evictTokens("invalidTokens"));
             assertThrows(CachingServiceClientException.class, () -> underTest.evictRules("invalidScopes"));
-        }
-
-    }
-
-    @ExtendWith(MockitoExtension.class)
-    @Nested
-    class CachingServiceAuthorization {
-
-        @Mock
-        private ApimlLogger apimlLogger;
-
-        @Test
-        void givenCredentials_whenSetCredentials_thenSetAuthorizationHeader() {
-            var service = new CachingServiceClient(mock(RestTemplate.class), null);
-            ReflectionTestUtils.setField(service, "cachingServiceUserId", "user");
-            ReflectionTestUtils.setField(service, "cachingServicePassword", "password");
-            ReflectionTestUtils.setField(service, "apimlLog", apimlLogger);
-
-            service.afterPropertiesSet();
-
-            // Verify that no warning is logged
-            verify(apimlLogger, times(0)).log("org.zowe.apiml.security.common.auth.missingDefaultCredentials");
-            // Verify that Basic authHeader is in the defaultHeaders map
-            var headers = (MultiValueMap<String, String>) ReflectionTestUtils.getField(service, "defaultHeaders");
-            assertEquals("Basic dXNlcjpwYXNzd29yZA==", headers.get(HttpHeaders.AUTHORIZATION).get(0));
-        }
-
-        @ParameterizedTest
-        @CsvSource({
-            ",password,,",
-            "user,,",
-            ",,"
-        })
-        void givenIncompleteCredentials_whenSetCredentials_thenDoNotSetAuthorization(String userId, String password) {
-            var service = new CachingServiceClient(mock(RestTemplate.class), null);
-            ReflectionTestUtils.setField(service, "cachingServiceUserId", userId);
-            ReflectionTestUtils.setField(service, "cachingServicePassword", password);
-            ReflectionTestUtils.setField(service, "apimlLog", apimlLogger);
-
-            service.afterPropertiesSet();
-
-            // Verify that warning is logged
-            verify(apimlLogger, times(1)).log("org.zowe.apiml.security.common.auth.missingDefaultCredentials");
-            // Verify that Basic authHeader is not in the defaultHeaders map
-            var headers = (MultiValueMap<String, String>) ReflectionTestUtils.getField(service, "defaultHeaders");
-            assertNull(headers.get(HttpHeaders.AUTHORIZATION));
         }
 
     }
